requirements.txt
SQLAlchemy~=1.3.20
marshmallow~=3.9.1
bcrypt~=3.2.0
PyJWT~=1.7.1
aiohttp~=3.7.3
aiopg~=1.0.0
alembic~=1.4.3
dependency_injector==4.5.1
nest_asyncio
aiohttp_jwt
pyyaml
aiohttp_cors